refactor: log messages and installation process

+++ install/desktop/app-brave.sh
@@ -1,10 +1,18 @@
 #!/bin/bash
 
-if [ ! -f /etc/apt/sources.list.d/brave-browser-release.list ]; then
-  [ -f /usr/share/keyrings/brave-browser-archive-keyring.gpg ] && sudo rm /usr/share/keyrings/brave-browser-archive-keyring.gpg
-  sudo curl -fsSLo /usr/share/keyrings/brave-browser-archive-keyring.gpg https://brave-browser-apt-release.s3.brave.com/brave-browser-archive-keyring.gpg
-  echo "deb [signed-by=/usr/share/keyrings/brave-browser-archive-keyring.gpg arch=amd64] https://brave-browser-apt-release.s3.brave.com/ stable main" | sudo tee /etc/apt/sources.list.d/brave-browser-release.list
-fi
+if ! command -v brave-browser &> /dev/null; then
+  print_header "Brave installation"
 
-sudo apt update
-sudo apt install -y brave-browser
+  if [ ! -f /etc/apt/sources.list.d/brave-browser-release.list ]; then
+    [ -f /usr/share/keyrings/brave-browser-archive-keyring.gpg ] && sudo rm /usr/share/keyrings/brave-browser-archive-keyring.gpg
+    sudo curl -fsSLo /usr/share/keyrings/brave-browser-archive-keyring.gpg https://brave-browser-apt-release.s3.brave.com/brave-browser-archive-keyring.gpg
+    echo "deb [signed-by=/usr/share/keyrings/brave-browser-archive-keyring.gpg arch=amd64] https://brave-browser-apt-release.s3.brave.com/ stable main" | sudo tee /etc/apt/sources.list.d/brave-browser-release.list
+  fi
+
+  sudo apt update
+  sudo apt install -y brave-browser
+
+  echo ""
+  log_success "Brave installed"
+  print_footer
+fi
