+++ install/desktop/optional/app-steam.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
-
 # Play games from https://store.steampowered.com/
-cd /tmp
-wget https://cdn.akamai.steamstatic.com/client/installer/steam.deb
-sudo apt install -y ./steam.deb
-rm steam.deb
-cd -
+
+if ! command -v steam &> /dev/null; then
+  print_header "Steam installation"
+
+  cd /tmp
+  wget https://cdn.akamai.steamstatic.com/client/installer/steam.deb
+  sudo apt install -y ./steam.deb
+  rm steam.deb
+  cd -
+
+  echo ""
+  log_success "Steam installed"
+  print_footer
+fi
